backend-main/requirements.txt
fastapi
uvicorn
python-dotenv
numpy
joblib
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
scikit-learn
